+++ code/header/basic/power_mismatch_struct.h
@@ -4,7 +4,7 @@
 struct POWER_MISMATCH_STRUCT
 {
     double greatest_mismatch_in_MVA;
-    size_t bus_with_greatest_mismatch;
+    unsigned int bus_with_greatest_mismatch;
 };
 
 #endif // POWER_MISMATCH_STRUCT_H
