+++ code/header/device/line.h
@@ -18,8 +18,8 @@ class LINE : public DEVICE
         LINE();
         virtual ~LINE();
 
-        void set_sending_side_bus(size_t bus);
-        void set_receiving_side_bus(size_t bus);
+        void set_sending_side_bus(unsigned int bus);
+        void set_receiving_side_bus(unsigned int bus);
         void set_identifier(string identifier);
         void set_sending_side_breaker_status(bool status);
         void set_receiving_side_breaker_status(bool status);
@@ -35,11 +35,11 @@ class LINE : public DEVICE
         void set_shunt_zero_sequence_y_at_receiving_side_in_pu(const complex<double>& y);
 
         void set_rating(RATING rating);
-        void set_meter_end_bus(size_t meter_bus);
+        void set_meter_end_bus(unsigned int meter_bus);
         void set_length(double length);
 
-        size_t get_sending_side_bus() const;
-        size_t get_receiving_side_bus() const;
+        unsigned int get_sending_side_bus() const;
+        unsigned int get_receiving_side_bus() const;
         BUS* get_sending_side_bus_pointer() const;
         BUS* get_receiving_side_bus_pointer() const;
 
@@ -57,25 +57,25 @@ class LINE : public DEVICE
         complex<double> get_shunt_zero_sequence_y_at_receiving_side_in_pu() const;
 
         RATING get_rating() const;
-        size_t get_meter_end_bus() const;
+        unsigned int get_meter_end_bus() const;
         double get_length() const;
 
         bool is_zero_impedance_line() const;
 
-        void set_fault(size_t to_bus, double location, const FAULT& fault);
-        double get_fault_location_of_fault(size_t index) const;
-        FAULT get_fault_at_location(size_t to_bus, double location) const;
+        void set_fault(unsigned int to_bus, double location, const FAULT& fault);
+        double get_fault_location_of_fault(unsigned int index) const;
+        FAULT get_fault_at_location(unsigned int to_bus, double location) const;
         void clear_all_faults();
-        void clear_fault_at_location(size_t to_bus, double location);
-        size_t get_fault_count() const;
+        void clear_fault_at_location(unsigned int to_bus, double location);
+        unsigned int get_fault_count() const;
         bool is_faulted() const;
 
         virtual bool is_valid() const;
         virtual void check();
         virtual void clear();
-        virtual bool is_connected_to_bus(size_t bus) const;
-        virtual bool is_in_area(size_t area) const;
-        virtual bool is_in_zone(size_t zone) const;
+        virtual bool is_connected_to_bus(unsigned int bus) const;
+        virtual bool is_in_area(unsigned int area) const;
+        virtual bool is_in_zone(unsigned int zone) const;
         virtual void report() const;
         virtual void save() const;
 
@@ -108,7 +108,7 @@ class LINE : public DEVICE
         complex<double> get_line_complex_apparent_impedance_at_receiving_side_in_ohm() const;
     private:
         //map<double,FAULT>::iterator get_iterator_of_fault_at_location_to_sending_side(double location);
-        size_t sending_side_bus, receiving_side_bus;
+        unsigned int sending_side_bus, receiving_side_bus;
         BUS* sending_side_busptr, *receiving_side_busptr;
         string   identifier;
         bool sending_side_breaker_status, receiving_side_breaker_status;
@@ -117,7 +117,7 @@ class LINE : public DEVICE
                         shunt_y_12_sending_side, shunt_y_0_sending_side,
                         shunt_y_12_receiving_side, shunt_y_0_receiving_side;
         RATING rating;
-        size_t meter_end_bus;
+        unsigned int meter_end_bus;
         double length;
 
         map<double,FAULT> faults;
